ldap: make it possible for administrators to impersonate another user
For verifying user's complaints about not being able to access a given
system – without having to ask for their password, because IT never asks
for your password.

This had to be done in the ldap module, because the current design
doesn't really allow for authentication and attribute retrieval as
separate steps. But this doesn't really matter for my case.